Added a live view script for 720p with calibration
The value in the strength of the "img.lens_corr(1.3)" was derived from just plugging in numbers and getting it as close as I could. This is because the OpenCV script that I used to find strength values was too complex with too many parameters for Open MV to run. 

Signed-off-by: justinvivadelli <justinvivadelli@gmail.com>